refactor: add `SelectArtworkListsForArtworkHeader`
