+++ src/app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/SelectArtworkListsForArtworkView.tsx
@@ -1,23 +1,15 @@
-import { Box, Button, Message, Spacer, quoteLeft, quoteRight } from "@artsy/palette-mobile"
 import { useArtworkListsContext } from "app/Components/ArtworkLists/ArtworkListsContext"
-import { ArtworkInfo } from "app/Components/ArtworkLists/components/ArtworkInfo"
 import { ArtworkListsBottomSheetSectionTitle } from "app/Components/ArtworkLists/components/ArtworkListsBottomSheetSectionTitle"
 import { AutomountedBottomSheetModal } from "app/Components/ArtworkLists/components/AutomountedBottomSheetModal"
 import { SelectArtworkListsForArtwork } from "app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/components/SelectArtworkListsForArtwork"
 import { SelectArtworkListsForArtworkFooter } from "app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/components/SelectArtworkListsForArtworkFooter"
+import { SelectArtworkListsForArtworkHeader } from "app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/components/SelectArtworkListsForArtworkHeader"
 import { ArtworkListsViewName } from "app/Components/ArtworkLists/views/constants"
 
 const SNAP_POINTS = ["50%", "95%"]
 
 export const SelectArtworkListsForArtworkView = () => {
-  const { state, dispatch, reset } = useArtworkListsContext()
-
-  const openCreateNewArtworkListView = () => {
-    dispatch({
-      type: "SET_CREATE_NEW_ARTWORK_LIST_VIEW_VISIBLE",
-      payload: true,
-    })
-  }
+  const { reset } = useArtworkListsContext()
 
   return (
     <AutomountedBottomSheetModal
@@ -30,33 +22,7 @@ export const SelectArtworkListsForArtworkView = () => {
         Select lists for this artwork
       </ArtworkListsBottomSheetSectionTitle>
 
-      <Box m={2}>
-        <ArtworkInfo artwork={state.artwork!} />
-
-        <Spacer y={2} />
-
-        <Button
-          variant="outline"
-          size="small"
-          width="100%"
-          block
-          onPress={openCreateNewArtworkListView}
-        >
-          Create New List
-        </Button>
-      </Box>
-
-      {!!state.recentlyAddedArtworkList && (
-        <>
-          <Message
-            variant="success"
-            title="List Created"
-            text={`Artwork will be added to ${quoteLeft}${state.recentlyAddedArtworkList.name}${quoteRight}`}
-          />
-          <Spacer y={1} />
-        </>
-      )}
-
+      <SelectArtworkListsForArtworkHeader />
       <SelectArtworkListsForArtwork />
       <SelectArtworkListsForArtworkFooter />
     </AutomountedBottomSheetModal>
